requirements.txt
Babel==2.6.0
certifi==2018.8.24
chardet==3.0.4
Django==1.11.15
django-phonenumber-field==2.0.0
django-uuidfield==0.5.0
djangorestframework==3.8.2
fcm-django==0.2.19
idna==2.7
phonenumberslite==8.9.13
pkg-resources==0.0.0
pyfcm==1.4.5
pytz==2018.5
requests==2.19.1
requests-toolbelt==0.8.0
urllib3==1.23
